+++ results/amziai.txt
@@ -1,2 +1,2 @@
-Žanui yra: 25 metų;
-Dovydui yra: 28 metų;
+👶 Žanui yra: 25 metų;
+🧑‍🎓 Dovydui yra: 28 metų;
